Added tests for SMS message entity.
SMS message entity and object now derive from the same test trait to ensure they behave the same.
Trait added in main src/Traits because namespace doesn't seem to be detected correctly when run via testing UI.
